local_ai_starter/requirements.txt
transformers
accelerate
bitsandbytes
sentence-transformers
faiss-cpu
jinja2
tqdm
python-dotenv
huggingface_hub
